seed moved to script; beginning component structure
