+++ script/seed.js
@@ -1,51 +1,445 @@
-/**
- * Welcome to the seed file! This seed file uses a newer language feature called...
- *
- *                  -=-= ASYNC...AWAIT -=-=
- *
- * Async-await is a joy to use! Read more about it in the MDN docs:
- *
- * https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Statements/async_function
- *
- * Now that you've got the main idea, check it out in practice below!
- */
-const db = require('../server/db')
-const {User} = require('../server/db/models')
+const {
+  db,
+  FoodItem,
+  GroceryList,
+  ListItem,
+  MacroGoal,
+  MicroGoal,
+  Supplement,
+  User,
+  Exercise,
+} = require('./db/models');
 
-async function seed () {
-  await db.sync({force: true})
-  console.log('db synced!')
-  // Whoa! Because we `await` the promise that db.sync returns, the next line will not be
-  // executed until that promise resolves!
+const foodItems = [
+  {
+    name: 'chicken breast',
+    servingSize: 5,
+    servingUnit: 'oz',
+    calories: 230,
+    protein: 43,
+    carbs: 0,
+    fat: 5
+  },
+  {
+    name: 'mac n cheese',
+    servingSize: 160,
+    servingUnit: 'g',
+    calories: 600,
+    protein: 20,
+    carbs: 100,
+    fat: 15
+  },
+  {
+    name: 'kale',
+    servingSize: 1,
+    servingUnit: 'cup',
+    calories: 33,
+    protein: 3,
+    carbs: 6,
+    fat: 0
+  },
+  {
+    name: 'bread',
+    servingSize: 1,
+    servingUnit: 'slice',
+    calories: 80,
+    protein: 3,
+    carbs: 17,
+    fat: 1
+  },
+  {
+    name: 'egg',
+    servingSize: 1,
+    servingUnit: 'large',
+    calories: 10,
+    protein: 6,
+    carbs: 0,
+    fat: 5
+  },
+  {
+    name: 'brown rice',
+    servingSize: 1,
+    servingUnit: 'cup',
+    calories: 248,
+    protein: 5,
+    carbs: 50,
+    fat: 2
+  },
+  {
+    name: 'black beans',
+    servingSize: 1,
+    servingUnit: 'cup',
+    calories: 227,
+    protein: 15,
+    carbs: 41,
+    fat: 1
+  },
+]
 
-  const users = await Promise.all([
-    User.create({email: 'cody@email.com', password: '123'}),
-    User.create({email: 'murphy@email.com', password: '123'})
-  ])
-  // Wowzers! We can even `await` on the right-hand side of the assignment operator
-  // and store the result that the promise resolves to in a variable! This is nice!
-  console.log(`seeded ${users.length} users`)
-  console.log(`seeded successfully`)
-}
+const supplements = [
+  {
+    name: 'milk thistle',
+    quantity: 100,
+    unitOfMeasure: 'mg',
+  },
+  {
+    name: 'magnesium',
+    quantity: 60,
+    unitOfMeasure: 'mg',
+  },
+  {
+    name: 'vegan b12',
+    quantity: 60,
+    unitOfMeasure: 'mg',
+  },
+  {
+    name: 'blueberry',
+    quantity: 60,
+    unitOfMeasure: 'mg',
+  },
+  {
+    name: 'cod liver-oil',
+    quantity: 8.4,
+    unitOfMeasure: 'mg',
+  },
+  {
+    name: 'calcium',
+    quantity: 60,
+    unitOfMeasure: 'mg',
+  },
+  {
+    name: 'zinc',
+    quantity: 60,
+    unitOfMeasure: 'mg',
+  },
+  {
+    name: 'calcium',
+    quantity: 60,
+    unitOfMeasure: 'mg',
+  },
+  {
+    name: 'selenium',
+    quantity: 60,
+    unitOfMeasure: 'mg',
+  },
+  {
+    name: 'iron',
+    quantity: 60,
+    unitOfMeasure: 'mg',
+  },
+  {
+    name: 'potassium',
+    quantity: 60,
+    unitOfMeasure: 'mg',
+  },
+];
 
-// Execute the `seed` function
-// `Async` functions always return a promise, so we can use `catch` to handle any errors
-// that might occur inside of `seed`
-seed()
-  .catch(err => {
-    console.error(err.message)
-    console.error(err.stack)
-    process.exitCode = 1
+const groceryLists = [
+  {
+    name: 'bulking grocery list',
+  },
+  {
+    name: 'cutting grocery list',
+  },
+  {
+    name: 'family grocery list',
+  },
+];
+
+const listItems = [
+  {
+    quantity: 8,
+    foodItemId: 1,
+    groceryListId: 3,
+  },
+  {
+    quantity: 1,
+    foodItemId: 2,
+    groceryListId: 3,
+  },
+  {
+    quantity: 12,
+    foodItemId: 5,
+    groceryListId: 3,
+  },
+]
+
+const macroGoals = [
+  {
+    name: 'Lean Bulking',
+    calories: 3000,
+    protein: 200,
+    carbs: 300,
+    fat: 200
+  },
+  {
+    name: 'Hard Bulking',
+    calories: 4000,
+    protein: 250,
+    carbs: 325,
+    fat: 225
+  },
+  {
+    name: 'Cutting',
+    calories: 2000,
+    protein: 200,
+    carbs: 200,
+    fat: 80
+  },
+  {
+    name: 'Maintenance',
+    calories: 2500,
+    protein: 200,
+    carbs: 250,
+    fat: 120
+  },
+  {
+    name: 'Keto',
+    calories: 2500,
+    protein: 250,
+    carbs: 25,
+    fat: 250
+  },
+]
+
+const microGoals = [
+  {
+    name: 'heart health',
+    vitaminA: 100,
+    vitaminB_12: 100,
+    vitaminC: 100,
+    vitaminD_3: 100,
+    vitaminK: 100,
+    potassium: 100,
+    sodium: 100,
+    magnesium: 100,
+    zinc: 100,
+  },
+];
+
+const exercises = [
+  {
+    name: 'Sprinting',
+    // category: Ambulatory,
+    // respiratory-type: Aerobic,
+    // resistance-type: Bodyweight
+  },
+  {
+    name: 'Jogging',
+    // category: Ambulatory,
+    // respiratory-type: Aerobic,
+    // resistance-type: Bodyweight
+  },
+  {
+    name: 'Running',
+    // category: Ambulatory,
+    // respiratory-type: Aerobic,
+    // resistance-type: Bodyweight
+  },
+  {
+    name: 'Walking',
+    // category: Ambulatory,
+    // respiratory-type: Aerobic,
+    // resistance-type: Bodyweight
+  },
+  {
+    name: 'Barbell Low Bar Squat',
+    // category: Strength/Hypertrophy,
+    // respiratory-type: Anaerobic,
+    // resistance-type: Weighted
+  },
+  {
+    name: 'Deadlift',
+    // category: Strength/Hypertrophy,
+    // respiratory-type: Anaerobic,
+    // resistance-type: Weighted
+  },
+  {
+    name: 'Romanian Deadlift',
+    // category: Strength/Hypertrophy,
+    // respiratory-type: Anaerobic,
+    // resistance-type: Weighted
+  },
+  {
+    name: 'Overhead Press',
+    // category: Strength/Hypertrophy,
+    // respiratory-type: Anaerobic,
+    // resistance-type: Weighted
+  },
+  {
+    name: 'Bench Press'
+    // category: Strength/Hypertrophy,
+    // respiratory-type: Anaerobic,
+    // resistance-type: Weighted
+  },
+  {
+    name: 'Barbell Row',
+    // category: Strength/Hypertrophy,
+    // respiratory-type: Anaerobic,
+    // resistance-type: Weighted
+  },
+  {
+    name: 'Barbell Lunge',
+    // category: Strength/Hypertrophy,
+    // respiratory-type: Anaerobic,
+    // resistance-type: Weighted
+  },
+  {
+    name: 'Barbell Split Squat',
+    // category: Strength/Hypertrophy,
+    // respiratory-type: Anaerobic,
+    // resistance-type: Weighted
+  },
+  {
+    name: 'Weighted Dip',
+    // category:  Strength/Hypertrophy,
+    // respiratory-type: Anaerobic,
+    // resistance-type: Weighted
+  },
+  {
+    name: 'Pullup',
+    // category:  Strength/Hypertrophy,
+    // respiratory-type: Anaerobic,
+    // resistance-type: Bodyweight
+  },
+  {
+    name: 'Chinup',
+    // category:  Strength/Hypertrophy,
+    // respiratory-type: Anaerobic,
+    // resistance-type: Bodyweight
+  },
+  {
+    name: 'Basic Pushup',
+    // category:  Strength/Hypertrophy,
+    // respiratory-type: Anaerobic,
+    // resistance-type: Bodyweight
+  },
+  {
+    name: 'Diamond Pushup',
+    // category:  Strength/Hypertrophy,
+    // respiratory-type: Anaerobic,
+    // resistance-type: Bodyweight
+  },
+  {
+    name: 'L-Sit',
+    // category:  Strength/Hypertrophy,
+    // respiratory-type: Anaerobic,
+    // resistance-type: Bodyweight
+  },
+  {
+    name: 'Ski Squat',
+    // category:  Strength/Hypertrophy,
+    // respiratory-type: Anaerobic,
+    // resistance-type: Bodyweight
+  },
+  {
+    name: 'Bodyweight Dip',
+    // category:  Strength/Hypertrophy,
+    // respiratory-type: Anaerobic,
+    // resistance-type: Bodyweight
+  },
+  {
+    name: 'Squat Jump',
+    // category:  Plyometric,
+    // respiratory-type: Anaerobic,
+    // resistance-type: Bodyweight
+  },
+  {
+    name: 'Split Squat Jump',
+    // category:  Plyometric,
+    // respiratory-type: Anaerobic,
+    // resistance-type: Bodyweight
+  },
+  {
+    name: 'Burpee',
+    // category:  Plyometric,
+    // respiratory-type: Anaerobic,
+    // resistance-type: Bodyweight
+  },
+  {
+    name: 'Depth Jump',
+    // category:  Plyometric,
+    // respiratory-type: Anaerobic,
+    // resistance-type: Bodyweight
+  },
+  {
+    name: 'Drop Jump',
+    // category:  Plyometric,
+    // respiratory-type: Anaerobic,
+    // resistance-type: Bodyweight
+  },
+  {
+    name: 'Slalom Jump',
+    // category:  Plyometric,
+    // respiratory-type: Anaerobic,
+    // resistance-type: Bodyweight
+  },
+  {
+    name: 'Standing Broad Jump',
+    // category:  Plyometric,
+    // respiratory-type: Anaerobic,
+    // resistance-type: Bodyweight
+  },
+  {
+    name: 'Bodyweight Lunge',
+    // category:  Plyometric,
+    // respiratory-type: Anaerobic,
+    // resistance-type: Bodyweight
+  },
+]
+
+db
+  .sync({ force: true })
+  .then(() => {
+    return Promise.all(
+      foodItems.map(foodItem => {
+        return FoodItem.create(foodItem);
+      })
+    );
   })
   .then(() => {
-    console.log('closing db connection')
-    db.close()
-    console.log('db connection closed')
+    return Promise.all(
+      supplements.map(supplement => {
+        return Supplement.create(supplement);
+      })
+    );
   })
-
-/*
- * note: everything outside of the async function is totally synchronous
- * The console.log below will occur before any of the logs that occur inside
- * of the async function
- */
-console.log('seeding...')
+  .then(() => {
+    return Promise.all(
+      groceryLists.map(groceryList => {
+        return GroceryList.create(groceryList);
+      })
+    );
+  })
+  .then(() => {
+    return Promise.all(
+      listItems.map(listItem => {
+        return ListItem.create(listItem);
+      })
+    );
+  })
+  .then(() => {
+    return Promise.all(
+      macroGoals.map(macroGoal => {
+        return MacroGoal.create(macroGoal);
+      })
+    );
+  })
+  .then(() => {
+    return Promise.all(
+      microGoals.map(microGoal => {
+        return MicroGoal.create(microGoal);
+      })
+    );
+  })
+  .then(() => {
+    return Promise.all(
+      exercises.map(exercise => {
+        return Exercise.create(exercise);
+      })
+    );
+  })
+  .then(() => {
+    console.log('success!!');
+  })
+  .catch(err => console.error(err));
